Separation finished.
