--- packages/minimongo/common_main.js
@@ -1,5 +1,0 @@
-import {LocalCollection} from './local_collection.js';
-import {Matcher} from './matcher.js';
-import {Sorter} from './sorter.js';
-
-Minimongo = {LocalCollection, Matcher, Sorter};
