+++ packages/minimongo/main.js
@@ -1,0 +1,6 @@
+import {LocalCollection as LocalCollection_} from './local_collection.js';
+import {Matcher} from './matcher.js';
+import {Sorter} from './sorter.js';
+
+Minimongo = {LocalCollection: LocalCollection_, Matcher, Sorter};
+LocalCollection = LocalCollection_;
