+++ packages/minimongo/sorter.js
@@ -1,3 +1,12 @@
+import {
+  ELEMENT_OPERATORS,
+  equalityElementMatcher,
+  expandArraysInBranches,
+  isOperatorObject,
+  makeLookupFunction,
+  regexpElementMatcher
+} from './common.js';
+
 // Give a sort spec, which can be in any of these forms:
 //   {"key1": 1, "key2": -1}
 //   [["key1", "asc"], ["key2", "desc"]]
